first draft for new bulk editing of lists. Allows handling of allow and block lists with an interface influenced by the manage list role
